requirements.txt
requests
praw
pytest
python-dotenv
